[TEST] user controller test코드 작성
