+++ src/modules/users/controllers/users.controller.ts
@@ -5,10 +5,6 @@ import { UsersService } from '../services/users.service';
 @Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
-  @Get()
-  getHello() {
-    return 'hello';
-  }
 
   @Post('/join')
   async joinUser(@Body() joinUserDto: JoinUserDto) {
